Switch to a flatter structure
Working on developer ergnonomics here; I want things to be as up-front
and easy-to-work-with as possible. I found I had a bunch of __init__.py
tabs open in my code editor with no easy way to tell which was which
implementation. The previous approach also made imports very concise,
but putting the entire implementation in __init__.py is not typical.